master/pyproject.toml
[project]
name = "master"
version = "0.1.0"
description = "Implementation of replicated log: master node"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.2",
    "python-dotenv>=1.0.1",
    "requests>=2.32.3",
    "sentry-sdk[fastapi]>=2.15.0",
    "uvicorn>=0.31.0",
]
